delete RectButton, rename RectButton2 to RectButton
